adstop: cache TLS configurations
Generating TLS configurations from hosts on the fly is very expensive on
ARM devices. Override the Connect handler to cache generated
configurations instead.

Note it still computes too much as concurrent requests for the same
configuration are not pooled together. This is addressed in following
changesets.

Also, the cache has no removal policy. This should do for now, lets see
how expensive it is to cache thousands of configurations first.