WIP adblock: separate optionless rules and build a massive regexp
I came back to the original adblock version, but this time the regular
expression captured nothing which means much better performances.
Unfortunately this is not the case, at least when running on a
RaspberryPi 2. Around 25k rules over 37k are compiled in the regular
expression but the matcher is globally a bit slower and consumes twice
the memory. The only advantage is it loads faster.