rules: combine leaf options with "or" not "and"
When several rules were sharing the same prefix, we tried to satisfy the
union of their options. Instead, try to satisfy one set of options.
Something like:

  ||foo.com^$domain=foo.com
  ||foo.com^$domain=bar.com

Is now treated like:

  ||foo,com^$domain=foo.com|bar.com

instead of:

  || foo.com^$domain=foo.com [and] bar.com

I do not know if this is specified anywhere but it makes much more sense
than the previous behaviour.

Reported by Ox Cart <ox.to.a.cart@gmail.com>.